fix: refactor zk proof service
