+++ src/background/services/zkProof/ZkProof.ts
@@ -1,0 +1,5 @@
+import ZkIdentityDecorater from "@src/background/identityDecorater";
+
+export abstract class ZkProof {
+  abstract genProof(identity: ZkIdentityDecorater, request: unknown): unknown;
+}
