+++ src/background/services/zkProof/__tests__/utils.test.ts
