+++ src/background/services/zkProof/index.ts
@@ -1,0 +1,3 @@
+export * from "./RLNProof";
+export * from "./SemaphoreProof";
+export * from "./utils";
