+++ src/background/services/zkProof/utils.ts
