+++ src/contentScripts/proof/__tests__/semaphore.test.ts
@@ -1,5 +1,5 @@
 import ZkIdentityDecorater from "@src/background/identityDecorater";
-import SemaphoreService from "@src/background/services/protocols/semaphore";
+import { SemaphoreProofService } from "@src/background/services/zkProof";
 
 import { ISemaphoreGenerateArgs, SemaphoreProofGenerator } from "..";
 
@@ -7,7 +7,7 @@ jest.mock("@src/background/identityDecorater", (): unknown => ({
   genFromSerialized: jest.fn(),
 }));
 
-jest.mock("@src/background/services/protocols/semaphore");
+jest.mock("@src/background/services/zkProof/SemaphoreProof");
 
 describe("contentScripts/proof/semaphore", () => {
   const defaultGenerateArgs: ISemaphoreGenerateArgs = {
@@ -38,7 +38,7 @@ describe("contentScripts/proof/semaphore", () => {
 
   test("should generate proof properly", async () => {
     const semaphoreProofGenerator = SemaphoreProofGenerator.getInstance();
-    const [semaphoreServiceInstance] = (SemaphoreService as jest.Mock).mock.instances as [{ genProof: jest.Mock }];
+    const [semaphoreServiceInstance] = (SemaphoreProofService as jest.Mock).mock.instances as [{ genProof: jest.Mock }];
     semaphoreServiceInstance.genProof.mockResolvedValue(emptyFullProof);
 
     const result = await semaphoreProofGenerator.generate(defaultGenerateArgs);
